requirements.txt
Django==1.5
Markdown==2.3.1
PIL==1.1.7
Pygments==1.6
django-appconf==0.6
django-compressor==1.3
django-datetime-widget==0.5
django-grappelli==2.4.4
django-guardian==1.1.0.beta
django-tinymce==1.5.1
gunicorn==0.17.4
hamlpy==0.82.2
six==1.3.0
wsgiref==0.1.2
